Create FluentMailer.java
method-chaining approach with lambda expressions: All the nonterminal methods return the instance. In addition, made the constructor private to disallow direct object creation. made the terminal method, send, a static method and it expects a Consumer as a parameter.
Rather than creating an instance, users will now invoke send and pass a block of code. The send method will create an instance, yield it to the block, and, upon return, complete any required validations and perform its final send operations.